Changed purple to the same one as on the save the date.
Changed nav bar to purple. When clicked on a link in the navbar, it
stays highlighted with the hover color.

Made font in the navbar and under Kendra and Vitaly same size which
forced to make the navbar wider which forced to make the popup sections
wider.

Resized the wedding party slider to make it narrower.